Create test for BA DAO

+++ V002DatabaseDesign/src/main/java/edu/bionic/sverkunov/com/DAODB3/classes/ReportResultBA.java
@@ -3,26 +3,31 @@ package edu.bionic.sverkunov.com.DAODB3.classes;
 
 public class ReportResultBA {
 
-	private int orderCountsOrderr;
+	private long orderCountsOrderr;
 	private double orderSumOrderrMenuItems;
 
-	public ReportResultBA(int orderCountsOrderr, double orderSumOrderrMenuItems) {
+	public ReportResultBA(long orderCountsOrderr, double orderSumOrderrMenuItems) {
 		this.orderCountsOrderr = orderCountsOrderr;
 		this.orderSumOrderrMenuItems = orderSumOrderrMenuItems;
 	}
 
-	public int getOrderCountsOrderr() {
+	public long getOrderCountsOrderr() {
 		return orderCountsOrderr;
 	}
-	public void setOrderCountsOrderr(int orderCountsOrderr) {
+
+	public void setOrderCountsOrderr(long orderCountsOrderr) {
 		this.orderCountsOrderr = orderCountsOrderr;
 	}
+
 	public double getOrderSumOrderrMenuItems() {
 		return orderSumOrderrMenuItems;
 	}
+
 	public void setOrderSumOrderrMenuItems(double orderSumOrderrMenuItems) {
 		this.orderSumOrderrMenuItems = orderSumOrderrMenuItems;
 	}
+
+	
 	
 	
 }
